requirements.txt
gunicorn==20.1.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
customtkinter==5.1.3
darkdetect==0.8.0
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.18.0
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
grpcio==1.54.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keyboard==0.13.5
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.6
packaging==23.1
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyenv-win==3.1.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQt6==6.5.0
PyQt6-Qt6==6.5.0
PyQt6-sip==13.5.1
pytesseract==0.3.10
requests==2.30.0
rsa==4.9
six==1.16.0
tqdm==4.65.0
urllib3==1.26.15
waitress==2.1.2
Werkzeug==2.3.4
yarl==1.9.2
